requirements.txt
# Core
Django==5.0.6
django-environ==0.11.2
gunicorn==22.0.0
psycopg2==2.9.9
django_tenants==3.6.1

# For the persistence stores
psycopg2-binary==2.9.9

# Rest apis
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
whitenoise==6.7.0
django-filter==24.2
